+ cacheService для предотвращения одних и тех же транзакций + тесты

+++ src/main/java/ru/war/robots/giftshop/service/ShopService.java
@@ -2,6 +2,7 @@ package ru.war.robots.giftshop.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.war.robots.giftshop.config.CacheService;
 import ru.war.robots.giftshop.config.InvalidDataException;
 import ru.war.robots.giftshop.jooq.tables.pojos.Item;
 import ru.war.robots.giftshop.jooq.tables.pojos.TUser;
@@ -14,6 +15,7 @@ import ru.war.robots.giftshop.repository.UserRepository;
 import ru.war.robots.giftshop.repository.WalletRepository;
 
 import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Stream;
 
 @Service
@@ -29,6 +31,8 @@ public class ShopService {
 
     @Autowired
     WalletRepository walletRepository;
+    
+    ConcurrentHashMap<Object, Boolean> cacheMap = new ConcurrentHashMap<>();
 
     public void buyItem(BuyDto dto) throws InvalidDataException {
         if (Stream.of(dto.getUserGuid(), dto.getItemId()).anyMatch(Objects::isNull)) {
@@ -46,17 +50,26 @@ public class ShopService {
 
         Wallet wallet = walletRepository.getByUserGuid(dto.getUserGuid());
         if (wallet == null
-                || wallet.getGoldCoins() < itemToBuy.getGoldPrice()
-                || wallet.getSilverCoins() < itemToBuy.getSilverPrice()
+            || wallet.getGoldCoins() < itemToBuy.getGoldPrice()
+            || wallet.getSilverCoins() < itemToBuy.getSilverPrice()
         ) {
             throw new InvalidDataException("User can't afford this item");
         }
-
-        inventoryRepository.addItem(user.getGuid(), itemToBuy.getId());
-
-        wallet.setSilverCoins(wallet.getSilverCoins() - itemToBuy.getSilverPrice());
-        wallet.setGoldCoins(wallet.getGoldCoins() - itemToBuy.getGoldPrice());
-        walletRepository.update(wallet);
+    
+        if (CacheService.get(dto) != null)  {
+            throw new InvalidDataException("Object already in transaction. Please wait for it to finish");
+        }
+        CacheService.put(dto);
+    
+        try {
+            inventoryRepository.addItem(user.getGuid(), itemToBuy.getId());
+    
+            wallet.setSilverCoins(wallet.getSilverCoins() - itemToBuy.getSilverPrice());
+            wallet.setGoldCoins(wallet.getGoldCoins() - itemToBuy.getGoldPrice());
+            walletRepository.update(wallet);
+        } finally {
+            CacheService.remove(dto);
+        }
     }
 
     public void gift(GiftDto dto) {
@@ -83,8 +96,17 @@ public class ShopService {
         }
 
         if (inventoryRepository.existsItemInInventory(dto.getSenderUserGuid(), dto.getItemId())) {
-            inventoryRepository.removeItem(dto.getSenderUserGuid(), dto.getItemId());
-            inventoryRepository.addItem(dto.getAcceptorUserGuid(), dto.getItemId());
+            if (CacheService.get(dto) != null)  {
+                throw new InvalidDataException("Object already in transaction. Please wait for it to finish");
+            }
+            CacheService.put(dto);
+            
+            try {
+                inventoryRepository.removeItem(dto.getSenderUserGuid(), dto.getItemId());
+                inventoryRepository.addItem(dto.getAcceptorUserGuid(), dto.getItemId());
+            } finally {
+                CacheService.remove(dto);
+            }
         } else {
             throw new InvalidDataException("Sender doesn't have this item");
         }
